+++ pyth-rng/src/command/generate.rs
@@ -6,12 +6,11 @@ use sha3::Digest;
 use crate::api::GetRandomValueResponse;
 
 use crate::config::GenerateOptions;
-
-use super::instantiate_contract_from_opts;
+use crate::ethereum::PythContract;
 
 /// Run the entire random number generation protocol to produce a random number.
 pub async fn generate(opts: &GenerateOptions) -> Result<(), Box<dyn Error>> {
-    let contract = Arc::new(instantiate_contract_from_opts(&opts.ethereum).await?);
+    let contract = Arc::new(PythContract::from_opts(&opts.ethereum).await?);
 
     let user_randomness = rand::random::<[u8; 32]>();
     let provider = opts.provider.parse::<Address>()?;
