+++ pyth-rng/src/command/request_randomness.rs
@@ -6,10 +6,10 @@ use sha3::Digest;
 
 use crate::config::RequestRandomnessOptions;
 
-use super::instantiate_contract_from_opts;
+use crate::ethereum::PythContract;
 
 pub async fn request_randomness(opts: &RequestRandomnessOptions) -> Result<(), Box<dyn Error>> {
-    let contract = Arc::new(instantiate_contract_from_opts(&opts.ethereum).await?);
+    let contract = Arc::new(PythContract::from_opts(&opts.ethereum).await?);
 
     let user_randomness = rand::random::<[u8; 32]>();
     let provider = opts.provider.parse::<Address>()?;
